refactor: create directorys for server and client side utils in utils directory, modify import/exports in components
Refactored the utils directory because a index file with exports of client AND server side code would throw errors while bundling. I splitted for that directory into server and client side files.
