+++ src/utils/client/index.tsx
@@ -1,0 +1,4 @@
+import { useLanguage, LanguageProvider } from './LanguageContext';
+import textMapFunction from './textMapFunction';
+
+export { useLanguage, LanguageProvider, textMapFunction };
