+++ src/utils/client/textMapFunction.tsx
